# managed/src/main/resources/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "ALPHA - NOT FOR EXTERNAL USE",
    "version" : "v1",
    "title" : "Yugabyte Platform APIs",
    "termsOfService" : "TODO(chirag)",
    "contact" : {
      "name" : "https://docs.yugabyte.com"
    },
    "license" : {
      "name" : "Polyform Free Trial License 1.0.0",
      "url" : "https://github.com/yugabyte/yugabyte-db/blob/master/licenses/POLYFORM-FREE-TRIAL-LICENSE-1.0.0.txt"
    }
  },
  "host" : "localhost",
  "basePath" : "/",
...
  "schemes" : [ "http", "https" ],
  "consumes" : [ "application/json" ],
  "produces" : [ "application/json" ],
  "paths" : {
...
    "/api/v1/customers/{cUUID}/universes/{uniUUID}/tables/{tableUUID}" : {
      "get" : {
        "tags" : [ "Tables" ],
        "summary" : "Describe table",
        "description" : "",
        "operationId" : "describe",
        "parameters" : [ {
          "name" : "cUUID",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        }, {
          "name" : "uniUUID",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        }, {
          "name" : "tableUUID",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "object",
              "additionalProperties" : {
                "type" : "object"
              }
            }
          }
        },
        "security" : [ {
          "apiKeyAuth" : [ ]
        } ]
      },
      "put" : {
        "tags" : [ "Tables" ],
        "summary" : "Alter table",
        "description" : "",
        "operationId" : "alter",
        "parameters" : [ {
          "name" : "cUUID",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        }, {
          "name" : "uniUUID",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        }, {
          "name" : "tableUUID",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "object",
              "additionalProperties" : {
                "type" : "object"
              }
            }
          }
        },
        "security" : [ {
          "apiKeyAuth" : [ ]
        } ]
      },
      "delete" : {
        "tags" : [ "Tables" ],
        "summary" : "Drop table",
        "description" : "",
        "operationId" : "drop",
        "parameters" : [ {
          "name" : "cUUID",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        }, {
          "name" : "uniUUID",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        }, {
          "name" : "tableUUID",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/YWTask"
            }
          }
        },
        "security" : [ {
          "apiKeyAuth" : [ ]
        } ]
      }
    },
    "/api/v1/customers/{cUUID}/universes/{uniUUID}/tables" : {
      "get" : {
        "tags" : [ "Tables" ],
        "summary" : "Universe list",
        "description" : "",
        "operationId" : "universeList",
        "parameters" : [ {
          "name" : "cUUID",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        }, {
          "name" : "uniUUID",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "type" : "object"
              }
            }
          }
        },
        "security" : [ {
          "apiKeyAuth" : [ ]
        } ]
      },
      "post" : {
        "tags" : [ "Tables" ],
        "summary" : "Create Table",
        "description" : "",
        "operationId" : "create",
        "parameters" : [ {
          "name" : "cUUID",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        }, {
          "name" : "uniUUID",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        }, {
          "in" : "body",
          "name" : "Table",
          "description" : "Table data to be created",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Universe task"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/YWTask"
            }
          }
        },
        "security" : [ {
          "apiKeyAuth" : [ ]
        } ]
      }
    },
    "/api/v1/customers/{cUUID}/universes/{uniUUID}/multi_table_backup" : {
      "put" : {
        "tags" : [ "Tables" ],
        "summary" : "Create Multiple table backup",
        "description" : "",
        "operationId" : "createMultiTableBackup",
        "parameters" : [ {
          "name" : "cUUID",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        }, {
          "name" : "uniUUID",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        }, {
          "in" : "body",
          "name" : "Table",
          "description" : "Table data to be created",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Multi table backup request params"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/YWTask"
            }
          }
        },
        "security" : [ {
          "apiKeyAuth" : [ ]
        } ]
      }
    },
    "/api/v1/metadata/column_types" : {
      "get" : {
        "tags" : [ "Tables" ],
        "summary" : "Get columns type",
        "description" : "",
        "operationId" : "getColumnTypes",
        "parameters" : [ ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "object",
              "additionalProperties" : {
                "type" : "object"
              }
            }
          }
        },
        "security" : [ {
          "apiKeyAuth" : [ ]
        } ]
      }
    },
    "/api/v1/customers/{cUUID}/universes/{uniUUID}/tables/{tableUUID}/bulk_import" : {
      "put" : {
        "tags" : [ "Tables" ],
        "summary" : "Bulk import table",
        "description" : "",
        "operationId" : "bulkImport",
        "parameters" : [ {
          "name" : "cUUID",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        }, {
          "name" : "uniUUID",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        }, {
          "name" : "tableUUID",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        }, {
          "in" : "body",
          "name" : "Bulk import",
          "description" : "Bulk data to be Imported",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Bulk import params"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/YWTask"
            }
          }
        },
        "security" : [ {
          "apiKeyAuth" : [ ]
        } ]
      }
    },
    "/api/v1/customers/{cUUID}/universes/{uniUUID}/tables/{tableUUID}/create_backup" : {
      "put" : {
        "tags" : [ "Tables" ],
        "summary" : "Create Multiple table backup",
        "description" : "",
        "operationId" : "createBackup",
        "parameters" : [ {
          "name" : "cUUID",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        }, {
          "name" : "uniUUID",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        }, {
          "name" : "tableUUID",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        }, {
          "in" : "body",
          "name" : "Backup",
          "description" : "Backup data to be created",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Backup Table Params"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/YWTask"
            }
          }
        },
        "security" : [ {
          "apiKeyAuth" : [ ]
        } ]
      }
    },
...
  },
  "securityDefinitions" : {
    "apiKeyAuth" : {
      "description" : "Api key passed as header",
      "type" : "apiKey",
      "name" : "X-AUTH-YW-API-TOKEN",
      "in" : "header"
    }
  },
  "definitions" : {
...
    "YWTask" : {
      "type" : "object",
      "required" : [ "taskUUID" ],
      "properties" : {
        "taskUUID" : {
          "type" : "string",
          "format" : "uuid"
        },
        "resourceUUID" : {
          "type" : "string",
          "format" : "uuid",
          "description" : "UUID of the resource being modified  by the task",
          "readOnly" : true
        }
      }
    },
...
  },
...
}